Remove the receive histogram used to derive rssi readings.  Instead use the rssi calculated (more efficiently) for use by the hal.  Fixup the last rx antenna setting which was previously in the rx histogram (we only ever used the last setting so just move it to the driver-private state).  This trims per-node storage and streamlines the rx path.
Note this change is not exactly compatible in that we may report old rssi
values and we only update the rssi based on rx'd data frames.  The previous
logic calculated rssi from all frames so for station mode it was updated with
data from beacon frames during idle periods.  We'll address this separately.

Reviewed by:	Kevin Yu