These changes do the following:
1. Enable use of QoS and short preamble in adhoc demo mode (discovered
neighbors are given the locally configured capabilities).  This is
especially useful for disabling ack's.
2. Time out nodes in ibss and adhoc demo modes.  This change may be
controversial especially in that madwifi has decided to ignore probe
request frames in adhoc demo mode (something I disagree with).
3. Change scangen bumping to pre-increment to be consistent with the bsd
code (you can eliminate the code to initialize nt_scangen to 1 as a
consequence; I have not done that).
4. Remove assertion in ieee80211_node_leave that verified the operating
mode was hostap, ibss, or ahdemo.  This call can occur when creating an
ahdemo vap, running it, then destroying it and moving to a sta mode vap.
In this scenario there may be a node in the table reclaimed through
the timeout mechanism and the assert in node leave will fire even though
it is safe to proceed because the node will never have an association id.

Signed-off-by: Sam Leffler <sam@errno.com>