Integrate ath_hal into ath
The ath_hal directory has no non-trivial code other than uudecode and
the makefiles.  Everything that is done there can be done in ath.

Also, ath_hal needed to be processed before ath to make sure that
opt_ah.h is created.  This cannot be ensured if all modules are built at
once by the kernel build system, which would be needed for Linux 2.6.17.

Now opt_ah.h is under version control, and the real include file name is
supplied by the Makefile.