Fix compilation for xscale-be-elf
Option -mshort-load-bytes is obsolete, so filter it out from COPTS after
the TARGET specific .inc file is included.

Machine specific compiler options should be generally left to the kernel
build system.  The options that come from HAL are currently preserved
just for one reason.  If they conflict with the options from the kernel,
it's a sign that incorrect TARGET may have been set and that the calling
conventions used in HAL may be incompatible with the kernel.

However, -mshort-load-bytes doesn't specify target architecture and
doesn't affect calling conventions.