Stop using ${...} notation in makefiles
There is no difference between $(...) and ${...} except that the later
is also supported by the shells.  Thus, when ${...} is used, it's not
immediately obvious that the expansion should be performed by make.  In
fact, some editors highlight $(...) and ${...} differently in makefiles.

Besides, both notations were used next to each other, suggesting that
there is so difference between them.

The ${...} notation is extremely rare in makefiles.  Linux uses it
mostly for shell expansion.