The data returned by an iwspy-get-ioctl (struct iw_quality) contains the field updated which informs if the quality data has been updated since the last iwspy-get-ioctl. This field is currently allways set to true, even if the quality information has not been updated since last ioctl.
The attached patch sets the field updated appropriately.

It works as follows:

The quality data returned by an iwspy-get-ioctl is based on the RSSI
value of the last recevied frame of a node (last-frame-rssi). When
reading the RSSI value, now additionally the last-frame-timestamp of
that node is stored. So if there is again an iwspy-get-ioctl, we can
check if a certain node has an updated RSSI value by comparing the
stored and current last-frame-timestamp of a node.

Signed-off-by: Christian Buennig <masala@web.de>