Check Linux version, not CRYPTO_ALG_ASYNC
Linux is firmly in the 2.6.19-rc phase, so whoever is using early
snapshots labeled as 2.6.18 with the new crypto API, should upgrade
right now.

It would be better to use crypto API version, but there is no such
thing, and CRYPTO_ALG_ASYNC doesn't seem to be a reliable test,
especially since it's not a feature we are using.