Avoid flexible arrays in structures for gcc-2.95 compatibility
Use zero-length arrays instead.  Flexible arrays are more standard, make
sizeof for the structure impossible (which is a good thing) and can be
initialized with a variable mumber of members.  But apparently we still
have users with gcc-2.95, and it's the only issue preventing MadWifi
compilation, so it would be better to fix it for now.

We could drop zero-size arrays once Linux 2.4 support is dropped.