Don't call the HAL gettsf64 function so often.
After profiling the driver to diagnose poor performance on
low-end systems (such as the Soekris 4526) it was found that
a large amount of time was being spent in ath_hal_gettsf64(). 

This was being caused by MadWiFi calling the function multiple
times unnecessarily, for example, for each transmitted packet.

This patch attempts to optimise the number of times we call
into ath_hal_gettsf64(), which reduces the number of PCI register
operations considerably and leads to a noticable performance
increase on low-end systems.