Don't lock irq twice in ath_beacon_generate()
Firstly, it is not needed.  Secondly, the implementation is fragile
because it reuses the same variable name for the flags.  Leaving the
"doubly protected" section early would leave interrupts disabled because
the inner value of __axq_lockflags would be used in both calls to
ATH_TXQ_UNLOCK_IRQ_EARLY.

Introduce macros ATH_TXQ_LOCK and ATH_TXQ_UNLOCK that don't affect IRQ
and use them for the inner lock.

Found by sparse.