this ports, with some changes, the IBSS merging work benoit has done in the dfs branch to trunk. especially notable is, that due to improved nexttbtt (next "target beacon transmit time") calculation we will now get the correct backoff behaviour for beacons, resulting in only one beacon per beacon interval (the current head version will send N beacons for N stations in the beacon interval because they are poorly synchronized). also because of the better timer synchronization there is no more time lag of up to 1 minute until we see beacons after a merge. thanks to benoit for figuring that out!
the difference between this patch and the version in the dfs-branch is that it 
still uses self linked descriptors and uses the SWBA interrupt only for 
updating the nexttbtt. this is necessary to recognize HW merges, for which we
don't get any notification by the hardware (see the thread "IBSS testing" on
this list for more details).

also i tried to clean up a bit, use a more descriptive function name for timer 
updates (ath_beacon_update_timers instead of ath_beacon_config) and generally 
better distingush between a HW merge and a SW merge.