This patch makes some debug messages about dropped frames a little more consistent and fixes an issue in hardstart where I was being too aggressive about requeing.
I thought the slab allocator would stop handing
out skbuff instances before the system was out
of free memory based upon some kind of limit on
the number of buffers, but this was nonsense.

Therefore, it is NOT a good idea to requeue
when skb allocation fails because this pretty
much basically means we are out of free memory.

Keeping the packets we do not have the RAM to
send and then stopping the transmit queue has
very bad side effects under low/zero memory 
conditions.  

Symptoms may include OOM killer and panics,
among other things.

It's better to drop tx packets when there's no
more RAM than to lock all the tx packets up
in the transmit queue.  Even if more skb
become freed up in the slab cache due to 
transmissions already in progress, evil things
still happen.