Implement linker set emulation
It's quite ugly, but it fixes linker errors.  The implementation uses
predefined sets of possible pointers declared as weak variables.  Those
variables are overridden by the actually present modules.

Interation is implemented using pointer tables pointing to the weak
variables.  This is done to avoid relying on the data layout in memory.