Don't try to determine ARCH automatically
ARCH was needed to calculate TARGET, but we don't need TARGET in
madwifi-free.  When cross-compiling the kernel, ARCH needs to be
specified anyway.  There is nothing wrong in requiring the same for
MadWifi.  Determining ARCH automatically is a nice feature, but it keeps
breaking for the new kernels.  Ideally, the kernel should keep ARCH in
.config rather than require it to be set every time.

Adjust documentation accordingly.